+++ src/types/properties.ts
@@ -1,3 +1,4 @@
+import { DataSourceRequestParams, DataSourceResponse } from './fetch';
 import { Entity, Field } from './model';
 
 export interface DynamicDataSource {
@@ -12,13 +13,25 @@ export interface DynamicDataSource {
   fields: Field[];
 }
 
-export type DesignDataSourceResult = {
+export type DataSourceFetcher = <T = DataSourceResponse>(requestParams?: DataSourceRequestParams) => Promise<T>;
+
+export type DataSourceResult = {
   entities: Entity[];
   fields: Field[];
   selectedEntity: Entity;
   selectedEntityCode: string;
+  fetchData: DataSourceFetcher;
 };
 
-export type UseDataSourceApi = (dataSource: DynamicDataSource) => DesignDataSourceResult;
+export type DataSourceSubscriber = (fetchData: DataSourceFetcher) => void;
+
+export type UseDataSourceApi = (dataSource: DynamicDataSource, subscriber?: DataSourceSubscriber) => DataSourceResult;
 
 export type ObjectVariableType = 'object' | 'objectList' | 'jsonSchema';
+
+export type UsePropertiesApi<P = any> = () => {
+  properties: P;
+  setProperties(props: P): void;
+  setPropertyByKey<K extends keyof P>(key: K, propValue: P[K]): void;
+  entities: Entity[];
+};